refactor: finish pattern-type migration and delete the re-export shims (#320)
