--- src/features/sequencer/types/pattern.ts
@@ -1,15 +1,0 @@
-/**
- * Pattern data model re-exports.
- *
- * The playback data model is owned by the audio engine
- * (core/audio/engine/pattern-types.ts); this shim preserves the historical
- * feature-layer import path for UI, store, and preset code.
- */
-
-export type {
-  TimingNudge,
-  StepSequence,
-  Voice,
-  VariationMetadata,
-  Pattern,
-} from "@/core/audio/engine/pattern-types";
