+++ src/test/fixtures.ts
@@ -4,14 +4,17 @@
  */
 
 import type { InstrumentRole } from "@/core/audio/engine/instrument/types";
+import type {
+  Pattern,
+  TimingNudge,
+  VariationId,
+} from "@/core/audio/engine/pattern-types";
 import { encodeWav } from "@/core/audio/export/wav-encoder";
 import type {
   InstrumentData,
   InstrumentParams,
 } from "@/features/instrument/types/instrument";
 import { createEmptyPattern } from "@/features/sequencer/lib/helpers";
-import type { Pattern, TimingNudge } from "@/features/sequencer/types/pattern";
-import type { VariationId } from "@/features/sequencer/types/sequencer";
 
 const FIXTURE_SAMPLE_RATE = 44100;
